+++ graphql/mutations/relations/deleteRelation.graphql
@@ -1,0 +1,5 @@
+mutation DeleteRelation($deleteData: DeleteRelationInput!) {
+  deleteRelation(deleteData: $deleteData) {
+    id
+  }
+}
