+++ package.json
@@ -27,7 +27,7 @@
     "test:e2e": "jest --forceExit --maxWorkers=3  --config ./test/config/jest.config.e2e.js",
     "test:it": "jest --forceExit --maxWorkers=4  --config ./test/config/jest.config.it.js",
     "test:auth": "jest --detectOpenHandles --runInBand --config ./test/config/jest.config.auth.js",
-    "test:auth-entities": "jest ---forceExit --maxWorkers=4 --config ./test/config/jest.config.auth-entities.js",
+    "test:auth-entities": "jest ---forceExit ---runInBand --config ./test/config/jest.config.auth-entities.js",
     "test:auth-global-roles": "jest ---forceExit --maxWorkers=4 --config ./test/config/jest.config.auth-global-roles.js",
     "test:ut": "jest --detectOpenHandles --runInBand --config ./test/config/jest.config.js",
     "test:nightly": "jest  --maxWorkers=6 --config ./test/config/jest.config.nightly.js",
