+++ test/non-functional/auth/my-privileges/global-roles/registered-user/challenge.it-spec.ts
@@ -8,7 +8,7 @@ import {
   getChallengeDataCodegen,
 } from '@test/functional-api/journey/challenge/challenge.request.params';
 import { deleteSpaceCodegen } from '@test/functional-api/journey/space/space.request.params';
-import { createRelation } from '@test/functional-api/integration/relations/relations.request.params';
+import { createRelationCodegen } from '@test/functional-api/relations/relations.request.params';
 import { createApplicationCodegen } from '@test/functional-api/user-management/application/application.request.params';
 import { entitiesId } from '@test/functional-api/zcommunications/communications-helper';
 import { TestUser } from '@test/utils';
@@ -89,7 +89,7 @@ beforeAll(async () => {
     TestUser.GLOBAL_ADMIN
   );
 
-  await createRelation(
+  await createRelationCodegen(
     entitiesId.challengeCollaborationId,
     'incoming',
     'relationDescription',
