+++ test/non-functional/auth/my-privileges/global-roles/registered-user/hub-private.it-spec.ts
@@ -1,28 +1,22 @@
 import {
   PostTypes,
   createPostOnCalloutCodegen,
-} from '@test/functional-api/integration/post/post.request.params';
+} from '@test/functional-api/callout/post/post.request.params';
 import {
   deleteSpaceCodegen,
   getSpaceDataCodegen,
-} from '@test/functional-api/integration/space/space.request.params';
-import { createRelation } from '@test/functional-api/integration/relations/relations.request.params';
+} from '@test/functional-api/journey/space/space.request.params';
+import { createRelationCodegen } from '@test/functional-api/relations/relations.request.params';
 import { createApplicationCodegen } from '@test/functional-api/user-management/application/application.request.params';
 import { entitiesId } from '@test/functional-api/zcommunications/communications-helper';
 import { TestUser } from '@test/utils';
-import { mutation } from '@test/utils/graphql.request';
 import { uniqueId } from '@test/utils/mutations/create-mutation';
-import {
-  changePreferenceSpace,
-  SpacePreferenceType,
-} from '@test/utils/mutations/preferences-mutation';
-import {
-  sendCommunityUpdate,
-  sendCommunityUpdateVariablesData,
-} from '@test/utils/mutations/update-mutation';
+import { changePreferenceSpaceCodegen } from '@test/utils/mutations/preferences-mutation';
 import { sorted__applyToCommunity_joinCommunity } from '../../common';
 import { createOrgAndSpaceCodegen } from '@test/utils/data-setup/entities';
 import { deleteOrganizationCodegen } from '@test/functional-api/organization/organization.request.params';
+import { SpacePreferenceType } from '@alkemio/client-lib';
+import { sendMessageToRoomCodegen } from '@test/functional-api/communications/communication.params';
 
 const organizationName = 'auth-ga-org-name' + uniqueId;
 const hostNameId = 'auth-ga-org-nameid' + uniqueId;
@@ -36,46 +30,37 @@ beforeAll(async () => {
     spaceName,
     spaceNameId
   );
-  await changePreferenceSpace(
+  await changePreferenceSpaceCodegen(
     entitiesId.spaceId,
-    SpacePreferenceType.ANONYMOUS_READ_ACCESS,
+    SpacePreferenceType.AuthorizationAnonymousReadAccess,
     'false'
   );
 
-  await changePreferenceSpace(
+  await changePreferenceSpaceCodegen(
     entitiesId.spaceId,
-    SpacePreferenceType.APPLICATIONS_FROM_ANYONE,
+    SpacePreferenceType.MembershipApplicationsFromAnyone,
     'true'
   );
-  await changePreferenceSpace(
+  await changePreferenceSpaceCodegen(
     entitiesId.spaceId,
-    SpacePreferenceType.JOIN_HUB_FROM_ANYONE,
+    SpacePreferenceType.MembershipJoinSpaceFromAnyone,
     'true'
   );
-  await changePreferenceSpace(
+  await changePreferenceSpaceCodegen(
     entitiesId.spaceId,
-    SpacePreferenceType.JOIN_HUB_FROM_HOST_ORGANIZATION_MEMBERS,
+    SpacePreferenceType.MembershipJoinSpaceFromHostOrganizationMembers,
     'true'
   );
 
   await createApplicationCodegen(entitiesId.spaceCommunityId, TestUser.QA_USER);
 
-  // await mutation(
-  //   createDiscussion,
-  //   createDiscussionVariablesData(
-  //     entitiesId.spaceCommunicationId,
-  //     DiscussionCategory.GENERAL,
-  //     'test'
-  //   )
-  // );
-
-  await mutation(
-    sendCommunityUpdate,
-    sendCommunityUpdateVariablesData(entitiesId.spaceUpdatesId, 'test'),
+  await sendMessageToRoomCodegen(
+    entitiesId.spaceUpdatesId,
+    'test',
     TestUser.GLOBAL_ADMIN
   );
 
-  await createRelation(
+  await createRelationCodegen(
     entitiesId.spaceCollaborationId,
     'incoming',
     'relationDescription',
@@ -103,7 +88,7 @@ describe('myPrivileges - Private Space', () => {
     // Act
     const response = await getSpaceDataCodegen(
       entitiesId.spaceId,
-      TestUser.NON_HUB_MEMBER
+      TestUser.QA_USER
     );
     const data = response.data?.space.authorization?.myPrivileges ?? [];
 
@@ -118,6 +103,11 @@ describe('myPrivileges - Private Space', () => {
         entitiesId.spaceId,
         TestUser.NON_HUB_MEMBER
       );
+      console.log('response', response.data);
+
+      console.log('response', response.data?.space);
+
+      console.log('response', response.data?.space.community?.authorization);
       const data =
         response.data?.space.community?.authorization?.myPrivileges ?? [];
