+++ test/non-functional/auth/my-privileges/global-roles/registered-user/opportunity-private-hub.it-spec.ts
@@ -8,7 +8,7 @@ import {
   deleteOpportunityCodegen,
   getOpportunityDataCodegen,
 } from '@test/functional-api/journey/opportunity/opportunity.request.params';
-import { createRelation } from '@test/functional-api/integration/relations/relations.request.params';
+import { createRelationCodegen } from '@test/functional-api/relations/relations.request.params';
 import { entitiesId } from '@test/functional-api/zcommunications/communications-helper';
 import { TestUser } from '@test/utils';
 import { uniqueId } from '@test/utils/mutations/create-mutation';
@@ -59,7 +59,7 @@ beforeAll(async () => {
     TestUser.GLOBAL_ADMIN
   );
 
-  await createRelation(
+  await createRelationCodegen(
     entitiesId.opportunityCollaborationId,
     'incoming',
     'relationDescription',
